Take percentiles in float64

+++ plugins/inputs/statsd/running_stats.go
@@ -99,7 +99,7 @@ func (rs *RunningStats) Count() int64 {
 	return rs.n
 }
 
-func (rs *RunningStats) Percentile(n int) float64 {
+func (rs *RunningStats) Percentile(n float64) float64 {
 	if n > 100 {
 		n = 100
 	}
@@ -109,7 +109,7 @@ func (rs *RunningStats) Percentile(n int) float64 {
 		rs.sorted = true
 	}
 
-	i := int(float64(len(rs.perc)) * float64(n) / float64(100))
+	i := int(float64(len(rs.perc)) * n / float64(100))
 	return rs.perc[clamp(i, 0, len(rs.perc)-1)]
 }
